+++ packages/react-foldkit/src/asyncData.ts
@@ -1,1 +1,1 @@
-export * from "../../../repos/foldkit/packages/foldkit/src/asyncData/public.js"
+export * from "foldkit/asyncData"