+++ packages/react-foldkit/src/message.ts
@@ -1,1 +1,1 @@
-export { m, m as make } from "../../../repos/foldkit/packages/foldkit/src/message/public.js"
+export { m, m as make } from "foldkit/message"