+++ packages/react-foldkit/src/schema.ts
@@ -1,1 +1,1 @@
-export * from "../../../repos/foldkit/packages/foldkit/src/schema/public.js"
+export * from "foldkit/schema"