+++ packages/react-foldkit/src/struct.ts
@@ -1,1 +1,1 @@
-export * from "../../../repos/foldkit/packages/foldkit/src/struct/public.js"
+export * from "foldkit/struct"