+++ packages/react-foldkit/src/update.ts
@@ -1,2 +1,2 @@
-export type * from "../../../repos/foldkit/packages/foldkit/src/update/public.js"
-export { combine, refresh } from "../../../repos/foldkit/packages/foldkit/src/update/public.js"
+export type * from "foldkit/update"
+export { combine, refresh } from "foldkit/update"